requirements.txt
astroid==2.8.2
azure-core==1.19.0
azure-storage-blob==12.9.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
cryptography==35.0.0
idna==3.3
isodate==0.6.0
isort==5.9.3
jedi==0.18.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
msrest==0.6.21
oauthlib==3.1.1
parso==0.8.2
platformdirs==2.4.0
pycparser==2.20
pylint==2.11.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
wrapt==1.12.1
